yolov5_deploy/requirements.txt
# certifi==2021.10.8
# charset-normalizer==2.0.12
# colorama==0.4.4
# cycler==0.11.0
# fonttools==4.29.1
# idna==3.3
# kiwisolver==1.3.2
# matplotlib==3.5.1
# numpy==1.22.2
# opencv-python==4.5.5.62
# packaging==21.3
# pandas==1.4.1
# Pillow==9.0.1
# pyparsing==3.0.7
# python-dateutil==2.8.2
# pytz==2021.3
# PyYAML==6.0
# requests==2.27.1
# scipy==1.8.0
# seaborn==0.11.2
# six==1.16.0
# torch==1.10.2+cu113
# torchaudio==0.10.2+cu113
# torchvision==0.11.3+cu113
# tqdm==4.63.0
# typing_extensions==4.1.1
# urllib3==1.26.8
# wget==3.2
# wincertstore==0.2


anyio==4.3.0
asttokens==2.4.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
docutils==0.21.2
easyAI==2.0.12
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
parso==0.8.3
prompt-toolkit==3.0.41
pure-eval==0.2.2
Pygments==2.17.2
pypiwin32==223
pypng==0.20220715.0
python-telegram-bot==21.1.1
pywin32==306
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
telegram==0.0.1
thread==2.0.5
torch==2.3.0
traitlets==5.14.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.12
Werkzeug==3.0.3
